requirements.txt
Django==1.6.1
django-haystack==2.1.0
requests==2.2.0
pyelasticsearch==0.6.1
